pyproject.toml
[project]
name = "text-to-sql-guardrails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.3",
    "fastapi>=0.136.0",
    "flower>=2.0.1",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.1.15",
    "openai>=2.32.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.13.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "structlog>=25.5.0",
    "uvicorn>=0.44.0",
]

[tool.pytest.ini_options]
markers = [
    # integration: requiere PostgreSQL + pgvector corriendo. Correr con: pytest -m integration
    "integration: tests que necesitan DB real y OPENAI_API_KEY",
]
